adds responsive react virtualized table
