+++ client/src/research/ResearcherDataPage.js
@@ -9,13 +9,11 @@ import logo from './logo.svg';
 import './ResearcherDataPage.css';
 import {hashInto, colorNames} from './Anonymize.js';
 
-import {Table, Column, AutoSizer, SortDirection, SortIndicator, CellMeasurer, CellMeasurerCache} from 'react-virtualized';
+import {AutoSizer} from 'react-virtualized';
 import 'react-virtualized/styles.css';
 
-// import styles from './ResearcherDataPage.css';
-
-// Substance of analysis
 import * as Analyses from './Analyses.js';
+import DynamicTable from './DynamicTable.js';
 
 
 /*
@@ -43,27 +41,6 @@ function rewriteAudioUrl(s3Folder, audioUrl) {
   const filename = audioUrl.slice(slashIndex + 1);
   return `${s3Folder}${filename}`;
 }
-
-function statsGroupedBy(allRows, keyName, keyFn) {
-  const rowGroups = _.groupBy(allRows, keyFn);
-  return _.map(rowGroups, (rows, keyValue) => {
-    const responses = rows.filter(row => row.type === 'on_response_submitted');
-    const textModeCount = responses.filter(row => row.json.textResponse).length;
-    const moveOnCount = responses.filter(row => !getAudioUrl(row)).length;
-    const anticipateCount = responses.filter(row => row.json.question.type === 'Anticipate').length;
-    const reflectCount = responses.filter(row => row.json.question.type === 'Reflect').length;
-
-    return {
-      [keyName]: keyValue,
-      anticipateCount,
-      moveOnCount,
-      textModeCount,
-      totalResponseCount: rows.length,
-      reflectCount
-    };
-  });
-}
-
 
 function hmtcaRowKey(row) {
   return [row.json.cohortKey, row.json.identifier, row.json.sessionId].join(" - ");
@@ -230,51 +207,6 @@ class Analysis extends Component {
     document.body.removeChild(a);
   }
 
-  _cache = new CellMeasurerCache({
-    fixedWidth: true,
-    minHeight: 25,
-  });
-
-  _getRowHeight ({ index }) {
-    return this._getDatum(index).size;
-  }
-
-  _headerRenderer({dataKey, sortBy, sortDirection, label}) {
-    return (
-      <div>
-        {label}
-        {sortBy === dataKey && <SortIndicator sortDirection={sortDirection} />}
-      </div>
-    );
-  }
-
-  _rowClassName({index}) {
-    if (index < 0) {
-      return "headerRow";
-    } else {
-      return index % 2 === 0 ? "evenRow virtualizedRow" : "oddRow virtualizedRow";
-    }
-  }
-
-  _wrappingCellRenderer = ({cellData, dataKey, parent, rowIndex}) => {
-    return (
-      <CellMeasurer
-        cache={this._cache}
-        columnIndex={0}
-        key={dataKey}
-        parent={parent}
-        rowIndex={rowIndex}>
-        <div
-          className={"tableColumn"}
-          style={{
-            whiteSpace: 'normal',
-          }}>
-          {cellData}
-        </div>
-      </CellMeasurer>
-    );
-  };
-
   onFetched(json) {
     const filtered = this.filter(json);
     this.setState({ json: filtered });
@@ -286,7 +218,6 @@ class Analysis extends Component {
 
   render() {
     const {json} = this.state;
-    console.log('rendering')
     return (
       <div className="Analysis">
         {(json)
@@ -298,9 +229,6 @@ class Analysis extends Component {
 
   renderJson(json) {
     const allRows = json.evidence.rows;
-    const statsForSessions = statsGroupedBy(allRows, 'sessionId', row => row.json.sessionId);
-    console.log(statsForSessions)
-    console.log(allRows)
 
     return (
       <div>
@@ -465,87 +393,20 @@ class Analysis extends Component {
   }
 
   renderEventsTableVirtualized(json) {
-    console.log(json);
-    console.log(json[0]);
-
     const {s3} = this.props.dataSet;
-
-    const disableHeader = false;
-    const headerHeight = 20;
-    const height = 400;
-    const hideIndexRow = false;
-    const overscanRowCount = 10;
-    const rowHeight = 40;
-    const rowCount = json.length;
-    const scrollToIndex = undefined;
-    const sortBy = 'index';
-    const sortDirection = "ASC";
-    const sortedList = json;
-    const useDynamicRowHeight = true;
-    const rowGetter=({index}) => json[index%json.length];
-
     return (
       <div>
         <AutoSizer disableHeight>
           {({width}) => (
-            <Table
-              deferredMeasurementCache={this._cache}
-              disableHeader={disableHeader}
-              headerClassName={"headerColumn"}
-              headerHeight={headerHeight}
-              height={height}
-              noRowsRenderer={this._noRowsRenderer}
-              overscanRowCount={overscanRowCount}
-              rowClassName={this._rowClassName}
-              rowHeight={rowHeight}
-              // rowHeight={this._cache.rowHeight}
-              // rowHeight={useDynamicRowHeight ? this._getRowHeight : rowHeight}
-              rowGetter={rowGetter}
-              rowCount={rowCount}
-              scrollToIndex={scrollToIndex}
-              sort={this._sort}
-              sortBy={sortBy}
-              sortDirection={sortDirection}
+            <DynamicTable
               width={width}
-            >
-              <Column 
-                dataKey="Timestamp" 
-                label="Timestamp"
-                disableSort = {false}
-                cellDataGetter={({ dataKey , rowData }) => moment(rowData[dataKey]).format('MM/DD/YY  h:mm:ssa')}
-                cellRenderer= {this._wrappingCellRenderer}
-                headerRenderer={this._headerRenderer}
-                width={300} 
-              />
-              <Column 
-                label="Email" 
-                dataKey="email" 
-                cellDataGetter={({ dataKey , rowData }) => rowData.json[dataKey]}
-                headerRenderer={this._headerRenderer}
-                width={400} 
-              />
-              <Column
-                label="Prompt"
-                dataKey="text"
-                cellDataGetter={({ dataKey , rowData }) => rowData.json.question.text || <div><span>Teacher Moments Scene: </span> <a href={"https://youtu.be/"+rowData.json.question.youTubeId}>https://youtu.be/{rowData.json.question.youTubeId}</a></div>}
-                cellRenderer= {this._wrappingCellRenderer}
-                headerRenderer={this._headerRenderer}
-                width={width}
-              />
-              <Column
-                label="Response"
-                dataKey="responseText"
-                cellDataGetter={({ dataKey , rowData }) => rowData.json[dataKey] || (getAudioUrl(rowData) && <audio controls={true} src={rewriteAudioUrl(s3, getAudioUrl(rowData))} /> )}
-                cellRenderer= {this._wrappingCellRenderer}
-                headerRenderer={this._headerRenderer}
-                width={width}
-              />
-            </Table>
+              list={json}
+              s3={s3}
+            />
           )}
         </AutoSizer>
       </div>
     );
-    // <tr key={row.id} >
   }
 }
